feat: allow toastService and modalService usage outside of components
