+++ src/composables/useToast.ts
@@ -1,7 +1,6 @@
-import { inject } from 'vue';
-import { InklineToastKey } from '@inkline/inkline/constants';
-import type { ToastService } from '@inkline/inkline/plugins';
+import { toastService } from '@inkline/inkline/services';
+import type { ToastService } from '@inkline/inkline/types';
 
-export function useToast() {
-    return inject(InklineToastKey) as ToastService;
+export function useToast(): ToastService {
+    return toastService;
 }
