+++ src/plugins/toast.ts
@@ -1,78 +1,22 @@
 import { createApp } from 'vue';
 import type { Plugin } from 'vue';
-import { createEventBus } from '@grozav/utils';
-import type { EventBus } from '@grozav/utils';
 import { defaultToastContainerId, InklineKey, InklineToastKey } from '@inkline/inkline/constants';
 import { IToastContainer } from '@inkline/inkline/components/IToastContainer';
 import { IconsPlugin } from '@inkline/inkline/plugins';
 import type { InklineService } from '@inkline/inkline/plugin';
-import type { StringOrRenderableType } from '@inkline/inkline/types';
-
-export type ToastPosition =
-    | 'top-left'
-    | 'top'
-    | 'top-right'
-    | 'right'
-    | 'bottom-right'
-    | 'bottom'
-    | 'bottom-left'
-    | 'left'
-    | string;
-
-export interface ToastOptions {
-    id: string;
-    title: StringOrRenderableType;
-    message: StringOrRenderableType;
-    icon: StringOrRenderableType;
-    position: ToastPosition;
-    duration: number;
-    dismissible: boolean;
-    showProgress: boolean;
-    color:
-        | 'light'
-        | 'dark'
-        | 'primary'
-        | 'secondary'
-        | 'info'
-        | 'success'
-        | 'warning'
-        | 'danger'
-        | string;
-
-    size: 'sm' | 'md' | 'lg';
-}
+import type { ToastOptions } from '@inkline/inkline/types';
+import { toastService } from '@inkline/inkline/services';
 
 export interface InklineToastOptions {
     toast?: Partial<Omit<ToastOptions, 'id'>>;
-}
-
-export interface ToastService {
-    show: (options: Partial<ToastOptions>) => void;
-    hide: (options: { id: string }) => void;
-    hideAll: () => void;
 }
 
 export interface InklineToastPluginOptions {
     inkline: InklineService;
 }
 
-export const toastEventBus: EventBus = createEventBus();
-
-export const createToastService = (): ToastService => ({
-    show: (options) => {
-        toastEventBus.emit('show', options);
-    },
-    hide: (options) => {
-        toastEventBus.emit('hide', options);
-    },
-    hideAll: () => {
-        toastEventBus.emit('hideAll', {});
-    }
-});
-
 export const ToastPlugin: Plugin = {
     install: (app, { inkline }: InklineToastPluginOptions) => {
-        const toastService = createToastService();
         const toastApp = createApp(IToastContainer);
 
         toastApp.provide(InklineKey, inkline);
@@ -80,6 +24,7 @@ export const ToastPlugin: Plugin = {
 
         if (typeof window !== 'undefined') {
             const containerId = defaultToastContainerId;
+            const containerDataAttribute = `data-${containerId}`;
             let container = document.querySelector(`#${containerId}`);
             if (!container) {
                 container = document.createElement('div');
@@ -87,7 +32,10 @@ export const ToastPlugin: Plugin = {
                 document.body.appendChild(container);
             }
 
-            toastApp.mount(container);
+            if (!container.hasAttribute(containerDataAttribute)) {
+                container.setAttribute(containerDataAttribute, '');
+                toastApp.mount(container);
+            }
         }
 
         app.config.globalProperties.$toast = toastService;
