+++ src/types/index.ts
@@ -2,5 +2,7 @@ export * from '@inkline/inkline/types/computed';
 export * from '@inkline/inkline/types/events';
 export * from '@inkline/inkline/types/guards';
 export * from '@inkline/inkline/types/manifest';
+export * from '@inkline/inkline/types/modal';
 export * from '@inkline/inkline/types/render';
+export * from '@inkline/inkline/types/toast';
 export * from '@inkline/inkline/types/validation';
